auth/requirements.txt
aiogram==3.15.0
fastapi==0.115.12
uvicorn==0.34.2
aiosqlite==0.21.0
python-dotenv==1.1.0
